fluffos-1.19

+++ src/program.c
@@ -6,7 +6,7 @@ int total_num_prog_blocks, total_prog_block_size;
 
 void reference_prog P2(const program_t *, progp, const char *, from)
 {
-    progp->ref++;
+    ((unsigned short)(progp->ref))++;
     debug(d_flag, ("reference_prog: /%s ref %d (%s)\n",
                progp->name, progp->ref, from));
 }
@@ -56,7 +56,7 @@ void deallocate_program P1(const program_t *, progp)
  */
 void free_prog P2(const program_t *, progp, int, free_sub_strings)
 {
-    progp->ref--;
+    ((unsigned short)(progp->ref))--;
     if (progp->ref > 0)
         return;
     if (progp->func_ref > 0)
